requirements.txt
# 更新后的requirements.txt，包含监控和配置相关依赖

# OKX API
okx-api==1.0.5

# 通用交易库
ccxt>=4.2.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 环境变量管理
python-dotenv>=1.0.0

# 定时任务
schedule>=1.2.0

# 日志记录
loguru>=0.7.0

# WebSocket
websockets>=12.0

# HTTP请求
requests>=2.31.0

# 加密
 cryptography>=41.0.0

# 异步支持
# asyncio  # Python3.4+内置库，不需要单独安装

# 终端颜色
 colorama>=0.4.6

# 表格显示
 tabulate>=0.9.0

# 图表
 matplotlib>=3.7.0
 plotly>=5.17.0

# 配置管理
 PyYAML>=6.0

# Web框架（用于监控仪表板）
 flask>=3.0.0
 flask-socketio>=5.3.0

# 数据库
# sqlite3  # Python内置库，不需要单独安装

# 系统监控
 psutil>=5.9.0